Fixed and finalized Friendly Creeper, changed how Ender Creeper teleport works and now has a 1 in 4 chance of teleporting those hit by the explosion, black listed bedrock in Reverse Creeper explosion and made it stop placing blocks below y-level 0,

+++ src/main/java/net/lomeli/ec/entity/addon/EntityMJCreeper.java
@@ -21,7 +21,7 @@ public class EntityMJCreeper extends EntityBaseCreeper {
 
     @Override
     public void explosion(int power, boolean flag) {
-        int radius = getPowered() ? (int) (this.explosionRadius * power) : this.explosionRadius;
+        int radius = getPowered() ? (this.explosionRadius * power) : this.explosionRadius;
         for (int x = -radius; x <= radius; x++)
             for (int y = -radius; y <= radius; y++)
                 for (int z = -radius; z <= radius; z++) {
